+++ src/includes/joedb/index_types.h
@@ -2,9 +2,10 @@
 #define joedb_index_types_declared
 
 #include <stdint.h>
+#include <stdlib.h>
 
 typedef uint16_t Table_Id;
 typedef uint16_t Field_Id;
-typedef uint64_t Record_Id;
+typedef size_t Record_Id;
 
 #endif
