+++ test/File_Test.cpp
@@ -164,7 +164,7 @@ TEST_F(File_Test, position_test)
  File file("new.tmp", Open_Mode::create_new);
  EXPECT_EQ(0ULL, file.get_position());
 
- file.set_position(uint64_t(-1));
+ file.set_position(size_t(-1));
  EXPECT_EQ(0ULL, file.get_position());
 
  const uint64_t N = 100;
